backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-filter>=23.3
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.54.0
dj-rest-auth>=4.0.1
google-auth>=2.22.0
requests>=2.31.0
assemblyai>=0.9.0
google-generativeai>=0.3.0
cloudinary>=1.36.0
yt-dlp>=2023.11.16 